used higher res depth inference and enhanced communication speed between api and host using buffers

+++ src/tests/shared_memory/client.py
@@ -1,7 +1,7 @@
 
 import numpy as np
 # Numpy Example
-f = np.memmap('raw_d_mm', dtype=np.float16, mode='r+', shape=(224,224))
+f = np.memmap('raw_d_mm', dtype=cfg.dtype, mode='r+', shape=(224,224))
 
 while True:
     print(f.var())
